+++ src/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 
 import products from './products';
 import product from './product';
-import cudCart from './cudCart';
+import addUpdateDeleteCart from './addUpdateDeleteCart';
 import carts from './carts';
 import cart from './cart';
 import snackbar from './snackbar';
@@ -10,7 +10,7 @@ import snackbar from './snackbar';
 export default combineReducers({
   products,
   product,
-  cudCart,
+  addUpdateDeleteCart,
   carts,
   cart,
   snackbar,
